requirements.txt
# BrainForge AI Knowledge Base - Requirements
# Generated from pyproject.toml for pip compatibility

# Core dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-core>=2.0.0

# Database
sqlalchemy>=2.0.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
pgvector>=0.2.0

# AI and ML
pydantic-ai>=0.1.0
openai>=2.8.1
numpy>=2.1.3

# PDF Processing
pypdf>=3.0.0
pdfplumber>=0.10.3

# Workflow and MCP
spiffworkflow>=1.0.0
fastmcp>=0.1.0

# Utilities
requests>=2.28.0
beautifulsoup4>=4.11.0
markdown>=3.4.0
python-multipart>=0.0.10
python-dotenv>=1.0.1
httpx>=0.28.1
python-magic>=0.4.27

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
ruff>=0.0.280
black>=23.0.0
mypy>=1.0.0
pre-commit>=3.0.0

# External API clients for researcher agent
google-api-python-client>=2.130.0
semanticscholar>=0.4.0
newsapi-python>=0.2.6

# Documentation
mkdocs>=1.6.0
mkdocs-material>=9.5.31

# Authentication
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4

# Caching and performance
redis>=7.1.0
cryptography>=42.0.0

# Monitoring and metrics
prometheus-client>=0.20.0
structlog>=23.0.0

# Backup and recovery
asyncpg>=0.29.0

# Production dependencies
gunicorn>=21.0.0
uvicorn[standard]>=0.23.0